Move all mushr translations to mushr_translation.hpp
